decomp another func and cleanup code
