+++ asm/code_8095014.s
@@ -5,41 +5,4 @@
 
 	.text
 
-	thumb_func_start sub_80951BC
-sub_80951BC:
-	push {r4,lr}
-	adds r4, r0, 0
-	bl FindOpenMailSlot
-	adds r2, r0, 0
-	movs r0, 0x1
-	negs r0, r0
-	cmp r2, r0
-	beq _080951F4
-	ldr r0, _080951F0
-	ldr r1, [r0]
-	lsls r0, r2, 1
-	adds r0, r2
-	lsls r0, 4
-	adds r1, r0
-	adds r0, r4, 0
-	ldm r0!, {r2-r4}
-	stm r1!, {r2-r4}
-	ldm r0!, {r2-r4}
-	stm r1!, {r2-r4}
-	ldm r0!, {r2-r4}
-	stm r1!, {r2-r4}
-	ldm r0!, {r2-r4}
-	stm r1!, {r2-r4}
-	movs r0, 0x1
-	b _080951F6
-	.align 2, 0
-_080951F0: .4byte gUnknown_203B480
-_080951F4:
-	movs r0, 0
-_080951F6:
-	pop {r4}
-	pop {r1}
-	bx r1
-	thumb_func_end sub_80951BC
-
         .align 2,0
