+++ include/save.h
@@ -49,10 +49,10 @@ struct UnkStruct_sub_8011DAC {
 struct UnkStruct_203B184 {
     /* 0x0 */ struct TeamInventory *MoneyItems;
     /* 0x4 */ struct unkStruct_203B45C *recruitedPokemon;
-    /* 0x8 */ u8 *unk8;
-    /* 0xC */ u8 *unkC;
+    /* 0x8 */ struct unkStruct_203B480 *unk8;
+    /* 0xC */ struct unkStruct_203B484 *unkC;
     /* 0x10 */ u32 *unk10;
-    /* 0x14 */ u32 *unk14;
+    /* 0x14 */ struct unkStruct_203B48C *unk14;
     /* 0x18 */ struct unkStruct_203B490 *mailInfo;
     /* 0x1C */ struct RescueTeamData *RescueTeamInfo;
     /* 0x20 */ u32 unk20;
