+++ src/save.c
@@ -76,10 +76,7 @@ EWRAM_DATA_2 struct QuickSaveWrite *gQuickSaveWrite = {0};
 
 extern s32 gUnknown_202DE28;
 extern bool8 *gFriendAreas;
-extern u8 *gUnknown_203B480;
-extern u8 *gUnknown_203B484;
 extern u32 *gUnknown_203B488;
-extern u32 *gUnknown_203B48C;
 extern u32 gUnknown_203B494;
 
 ALIGNED(4) const char sGameInternalVersion[] = _("POKE_DUNGEON__05");
@@ -155,10 +152,6 @@ extern u32 RestoreRecruitedPokemon(void* a, s32 b);
 extern u32 SavePokemonStruct2(u8 *, u32);
 extern u32 RestorePokemonStruct2(void* a, s32 b);
 extern u32 sub_80921C4(u8 *, u32);
-extern u8 *sub_8095100(void);
-extern u32 *sub_8095108(void);
-extern u32 *sub_8095110(void);
-extern void sub_8095118(void);
 extern u32 sub_80954CC(void* a, s32 b);
 extern u32 sub_8095624(u8 *, u32);
 extern u32 RestoreMailInfo(void* a, s32 b);
@@ -168,7 +161,6 @@ extern u32 sub_8097D60(u8 *, u32);
 extern u32 sub_8097D98(void* a, s32 b);
 extern void sub_80993E4();
 extern void sub_800135C(void);
-extern u8 *sub_80950F8(void);
 extern void sub_80958E4(u32 *a, u32 b);
 extern struct unkStruct_203B490 *GetMailInfo(void);
 extern void InitializeMailJobsNews(void);
